Finish migration to redux-observables for shared state syncing
Refactor the shared-worker state syncing API to use redux observables on
the worker-side. Actions are forwarded to the clients with rootState put
in action metadata so that actions are logged in Redux DevTools just as
if the store was running locally (previously DevTools showed just a
series of 'root/sync' actions).