Refactor approach to syncing so whole state is not replaced every update
Being able to detect which parts of state have changed and which haven't
will be important for several aspects of this app so that downstream
updates to UI, the audio engine, etc only trigger when relevant changes
have occurred. This ability was broken by the previous approach to
syncing, which replaced the entire app state on every update from the
shared worker state.

The previous paradigm has been replaced by one that uses the worker
state only for initialization, after which state changes are applied by
the same reducers that are used on the worker. So long as no actions are
lost or received out-of-order, state should remain in sync between the
worker and the local clients. (If going out-of-sync does become an issue,
copies of the entire worker state could be applied to local clients on a
periodic basis, similar to how keyframes are used in video codecs.)