Create new "instruments" state slice and "Engine" classes
Create a new "engine" folder hierarchy that facilitates instantiating
and connecting sequences to instruments based on app state. An abstract
"EngineInstruments" class encapsulates the instrument API, and a
top-level "Engine" class provides a high-level interface for adding,
updating, and connecting sequences and instruments. Sequence and
Instrument epics observe state changes and call update functions
defined on the engine.

A DrumMachine class implements the EngineInstrument API and is hooked
up to receive triggers, but is currently just logging them and throwing
them away.