Add "currentStep" to Sequence state to enable highlighting current step
This commit implements highlighting of the active step in the sequence
views, but it has also introduced performance problems, with the UI
updates losing their smoothness and some audio glitches becoming
noticable. Blacklisting the "setCurrentStep" action from syncing with
the shared store worker only slightly improved things. Performance
profiling will be required to determine if the primary cause of the
issues is in the updating of the UI components, the state handling, or
both.