Hack in a p5-backed video instrument
Create a VideoOutput UI component that runs a p5 sketch defined by a
test "Circle" EngineInstrument, and hook up the component so that the
Circle instrument receives triggers from the same sequencer as the
Casio. To demonstrate that triggers are working and synced with the
audio, Circle updates the radius of a circle based on the frequency of
the trigger.