List sequence outputs, add warning when instruments are disabled
Create some new selectors in apiSequences to enable the listing of
sequence outputs to the left of the sequencer grids. Also, add a warning
when audio outputs are disabled to help the user get out of a state
where they lost audio output because they closed the first tab they
opened (once the first-opened tab is closed it is currently impossible
to get audio output back until you close all tabs and reload the page).