Create a websocket API and go server for syncing clocks
Create a simple go server with a websocket API that can be used by
clients to sync their clocks. The clients use rolling averages of
computed offsets from the server clock (first and second derivatives) to
determine corrections to be applied to the client's local clock. Tests
running on a local network result in average error of less than ±0.1ms.