Tear out most of Minor Assignment 1, replace with bare-bones p5 sketch
Create a p5 sketch with inputs for user name, active instrument, and
input device, and use the websocket connection to send/receive updates
for all connected users. Midi input is connected to Tone, but not yet
running through the server.

On the server side, create an event bus and complementary types to
facilitate distribution of user updates to all connected clients.