Fix a bug that caused clocks not to sync
A bug in WSClock caused the sync loop to be canceled if the socket
connection was not open when the WSClock was created; a quick fix is
implemented to re-create the WSClock upon opening of the websocket
connection.

Also, a second piano instrument is added so that multiple users can play
piano without accidentally releasing each other's notes, which is caused
by a bug (or unexpected behavior) in Tone where calling 'triggerRelease'
cancels future attacks that have been previously scheduled.