Add movement controls (arrow keys) for 3D camera
Extend p5.easycam with a "panGround" function that moves the camera
parallel to the ground by projecting screen-space vectors onto the
ground plane. Connect arrow-key presses to an accelleration state for
the user, and use the accelleration to compute velocity every draw and
update the camera with the new panGround function. Draw a sphere to show
the camera's center location.