Begin reimplementing avatar in a simple, component-based 3D engine
Create a very simple component-based engine inspired by Unity. Objs
contain a transform, a custom draw function, and components. The
'Physical' component can be added to an Obj to make it respond to
forces and be constrained to hard-coded world bounds.

The new engine is used to add a sphere to the scene that roughly
mimics the behavior of the previous EasyCam-based avatar.