Sync user avatar transforms over the websocket connection
Extend the server's websocket api to support new UserXform data,
and use the new api to enable clients to send and receive updates
to other users' avatars. Update VisualNotes to make the note shapes
follow the avatar of the user that played the notes.

Also, solve a bug in draw2D that caused avatar labels to appear
on-screen for objects that were behind the camera.