Extend Loops UI to support multiple loops
The 'Active' loop is displayed next to the metronome dial, other loops
(if any) appear smaller below. Click on an other loop to make it active,
Shift+Click to delete. 'M' key mutes the active loop.

New buttons in the bottom right corner hide/show the metronome dial and
the loops.