Add Dancer instrument to give avatar a dance move (the 'dip')
To support the new Dancer instrument, the Obj class is extended to
include child objects that inherit the parent Obj's transform. The
new DancerObj is now responsible for drawing the avatar, and its
DanceMoves components uses a Tone.Envelope to implement the 'dip'
move by dropping the DancerObj's Y position based on the envelope
value.