Create a BlackHole instrument backed by a noisy FM Tone synth chain
BlackHole is similar to the polySynth in that it creates Objs on note
attacks, but it is backed by a Tone synth with frequency modulated by
noise — the modwheel affects the gain of the noise modulation and
the twistiness of the blackhole objs; velocity affects the opacity of
the objs. Fractally goodness.